Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main(){
-    int i,j,k;
-    scanf("%d",&i);
-    scanf("%d",&j);
-    scanf("%d",&k);
+    float i,j,k;
+    scanf("%f",&i);
+    scanf("%f",&j);
+    scanf("%f",&k);
     float ans=(i+j+k)/3;
     printf("Average: %.2f",ans);
 }
